[BUGFIX] Handle VIP guests in `guestlist:list-loaded` event handler
Previously, the CSS classes of rows representing VIP guests were changed
in a global, low level, place being a no-op in most cases.

The event handler for `guestlist:list-loaded` is moved into a
separate function which allows dispatching the event on the guestlist
table only. Finally, the VIP handling is now done in that extracted event
handler.