[TASK] Change default sorting and add tablesort library
The default sorting of guests is changed to firstName=ASC, lastName=ASC,
the check in date is not taken into account anymore, which renders the
manual table sorting being applied after updating a guest obsolete and is
removed.

The library `tablesort` is added to enable sorting the table by clicking
a table header, which is enabled for "first name" and "last name".

Fixes: #19